Refactor move connectToDb.js
